refactor: rename submodules of Filterable
Rename submodules of Filterable so that they fall within the
WcmcComponents::Filterable namespace. Remove Table prefix from each
